add a useless static_cast to make the complaintion shut up

+++ test/lru_cache.cc
@@ -9,7 +9,7 @@ namespace
 int func(int i, unsigned b, double d)
 {
     ++evaluate_cnt;
-    return i + b + d;
+    return static_cast<int>(i + b + d);
 }
 
 TEST(lru_cache, basic)
